+++ app/lab/murajaah-at-taisir/page.tsx
@@ -1,0 +1,44 @@
+import { Metadata } from "next";
+import { notFound } from "next/navigation";
+
+import { SITE_NAME } from "@/app/constants";
+import { fetchQuranPage } from "./lib/fetchQuranPage";
+import { QuranPageRenderer } from "./components/QuranPageRenderer";
+
+type PageProps = {
+  searchParams: Promise<{ [key: string]: string | string[] | undefined }>;
+};
+
+const MIN_PAGE = 1;
+const MAX_PAGE = 604;
+
+export const metadata: Metadata = {
+  title: "Aplikasi Murajaah at Taisir | HiRahmat.Dev",
+  description:
+    "Terinspirasi dari buku Muraja'ah At-Taisir (30 Hari Hafal Al-Qur'an) - Adi Hidayat",
+  openGraph: {
+    title: "Aplikasi Murajaah at Taisir | HiRahmat.Dev",
+    description:
+      "Terinspirasi dari buku Muraja'ah At-Taisir (30 Hari Hafal Al-Qur'an) - Adi Hidayat",
+    siteName: SITE_NAME,
+    url: "/lab/murajaah-at-taisir",
+    images: "/lab/murajaah-at-taisir/opengraph-image.png",
+    locale: "id",
+    type: "website",
+  },
+};
+
+export default async function MurajaahAtTaisirPage({
+  searchParams,
+}: PageProps) {
+  const page = Number((await searchParams).page || "1");
+
+  if (!page || isNaN(page) || page > MAX_PAGE || page < MIN_PAGE) {
+    notFound();
+  }
+
+  const isRightPage = page % 2 === 1;
+  const pageN = await fetchQuranPage(page);
+
+  return <QuranPageRenderer pageN={pageN} isRightPage={isRightPage} />;
+}